+++ @xen-orchestra/rest-api/src/helpers/helper.type.mts
